+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.0] - 2026-09-22
+
+### Added
+- Added language-server completion for text links.
+- Added language-server support for renaming text nodes and their links.
+
+### Changed
+- Rendered language-server node previews as Markdown and made them available from node titles.
+
 ## [0.11.1] - 2026-09-22
 
 ### Changed